pyproject.toml
[tool.poetry]
name = "gpt-pilot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pilot"}]

[tool.poetry.dependencies]
python = "^3.11"
blessed = "^1.20.0"
certifi = "^2023.5.7"
charset-normalizer = "^3.2.0"
colorama = "^0.4.6"
distro = "^1.8.0"
idna = "^3.4"
jsonschema = "^4.19.1"
Jinja2 = "^3.1.2"
MarkupSafe = "^2.1.3"
peewee = "^3.16.2"
prompt-toolkit = "^3.0.39"
psutil = "^5.9.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
python-editor = "^1.0.4"
pytest = "^7.4.2"
pyyaml = "^6.0.1"
questionary = "^1.10.0"
readchar = "^4.0.5"
regex = "^2023.6.3"
requests = "^2.31.0"
six = "^1.16.0"
termcolor = "^2.3.0"
tiktoken = "^0.4.0"
urllib3 = "^1.26.6"
wcwidth = "^0.2.6"
yaspin = "^2.4.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
